+++ packages/react-tinacms-editor/src/tinacms-inline/inline-wysiwyg.tsx
@@ -18,28 +18,69 @@ limitations under the License.
 
 import * as React from 'react'
 import { useCMS } from 'tinacms'
-import { InlineField } from 'react-tinacms-inline'
+import { InlineField, FocusRing } from 'react-tinacms-inline'
+import { FocusRingOptions } from 'react-tinacms-inline/src/styles'
 import { Wysiwyg } from '../components/Wysiwyg'
-import { EditorProps } from '../types'
+import { EditorProps, ImageProps } from '../types'
 
 export interface InlineWysiwygFieldProps extends Omit<EditorProps, 'input'> {
   name: string
   children: any
+  focusRing?: boolean | FocusRingOptions
 }
 
 export function InlineWysiwyg({
   name,
   children,
+  focusRing = true,
+  imageProps: passedInImageProps,
   ...wysiwygProps
 }: InlineWysiwygFieldProps) {
   const cms = useCMS()
+
+  const imageProps: ImageProps | undefined = React.useMemo(() => {
+    if (!passedInImageProps) return
+    return {
+      async upload(files: File[]) {
+        const filesToUpload = files.map(file => ({
+          directory: passedInImageProps?.directory || '',
+          file,
+        }))
+
+        const allMedia = await cms.media.store.persist(filesToUpload)
+
+        return allMedia.map(media => {
+          if (passedInImageProps.parse) {
+            return passedInImageProps.parse(media.filename)
+          } else {
+            return media.filename
+          }
+        })
+      },
+      previewSrc(src) {
+        return cms.media.store.previewSrc(src)
+      },
+      ...passedInImageProps,
+    }
+  }, [
+    cms.media.store,
+    passedInImageProps?.directory,
+    passedInImageProps?.previewSrc,
+    passedInImageProps?.upload,
+  ])
+
+  if (cms.disabled) {
+    return children
+  }
+
   return (
     <InlineField name={name}>
       {({ input }: any) => {
-        if (cms.enabled) {
-          return <Wysiwyg input={input} {...wysiwygProps} />
-        }
-        return <>{children}</>
+        return (
+          <FocusRing name={name} options={focusRing}>
+            <Wysiwyg input={input} {...wysiwygProps} imageProps={imageProps} />
+          </FocusRing>
+        )
       }}
     </InlineField>
   )