+++ packages/react-tinacms-editor/src/types.ts
@@ -25,8 +25,10 @@ import { Schema } from 'prosemirror-model'
 import { Format } from './translator'
 
 export interface ImageProps {
+  parse(filename: string): string
+  directory?: string
   upload?: (files: File[]) => Promise<string[]>
-  previewUrl?: (url: string) => string
+  previewSrc?: (url: string) => string | Promise<string>
 }
 
 export interface KeymapPlugin {