+++ packages/@tinacms/teams/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.26.0](https://github.com/tinacms/tinacms/compare/v0.25.0...v0.26.0) (2020-08-03)
+
+**Note:** Version bump only for package @tinacms/teams
+
+
+
+
+
 ## [0.2.2](https://github.com/tinacms/tinacms/compare/@tinacms/teams@0.2.2-alpha.1...@tinacms/teams@0.2.2) (2020-04-27)
 
 **Note:** Version bump only for package @tinacms/teams